refactor: extract seatbelt profile adapter
